backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Network Scanning
python-nmap==0.7.1
scapy==2.5.0

# WebSocket
websockets==12.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0

# Export functionality
jinja2==3.1.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
